Add regression test for VM globalThis.eval fix
Adds comprehensive test covering:
- Basic globalThis.eval access in VM contexts
- happy-dom's exact failing pattern
- DONT_CONTEXTIFY mode behavior

Also removes incorrectly placed test file.
